Move old cuts to RetiredCuts (#44)
RetiredCuts doesn't compile. It's just there for reference.

* Missed a couple of cuts in GetCutName. Done now.
