+++ includes/Constants.h
@@ -21,10 +21,7 @@ enum ECuts {
   kVtx,
   kMinosMatch,
   kMinosCharge,
-  kMinosCoil,
   kMinosMuon,
-  kThetaMu,
-  kDeadTime,
   kNProngs,
   kWexp,
   kNPionCandidates,
@@ -32,18 +29,8 @@ enum ECuts {
   kAtLeastOnePionCandidateTrack,
   kNode,
   kPionMult,
-  kAtLeastOneAnchoredProng,
-  kNoInteractionVertexMichels,
   kAtLeastOneMichel,
-  kAtLeastOneBrandonMichel,
-  kExactlyOneEndpointMichel,
-  kOldMichel,
-  kdEdx,
-  kAtLeastOneLLRCandidate,
-  kAtLeastOneNodeCandidate,
   kLLR,
-  kIsoBlobs,
-  kIsoProngSep,
   kIsoProngs,
   kPmu,
   kAllCuts,
