+++ includes/Cuts.cxx
@@ -232,30 +232,14 @@ bool PassesCut(const CVUniverse& univ, const ECuts cut, const bool is_mc,
     case kMinosCharge:
       return MinosChargeCut(univ);
 
-    case kMinosCoil:
-      return MinosCoilCut(univ);
-
     case kMinosMuon:
       return MinosMatchCut(univ) && MinosChargeCut(univ);
-      //&& MinosCoilCut(univ);
-
-    case kThetaMu:
-      return ThetamuCut(univ);
-
-    case kDeadTime:
-      return DeadTimeCut(univ);
 
     case kWexp:
       return WexpCut(univ, signal_definition);
 
-    case kIsoBlobs:
-      return IsoBlobCut(univ);
-
     case kIsoProngs:
       return IsoProngCut(univ);
-
-    case kIsoProngSep:
-      return IsoProngSepCut(univ);
 
     case kPmu:
       return PmuCut(univ);
@@ -277,18 +261,6 @@ bool PassesCut(const CVUniverse& univ, const ECuts cut, const bool is_mc,
 
     case kAtLeastOnePionCandidateTrack:
       return GetQualityPionCandidateIndices(univ).size() > 0;
-
-    case kAtLeastOneLLRCandidate:
-      return AtLeastOneLLRCandidateCut(univ);
-
-    case kAtLeastOneNodeCandidate:
-      return AtLeastOneNodeCandidateCut(univ);
-
-    case kAtLeastOneAnchoredProng:
-      return AtLeastOneAnchoredProngCut(univ);
-
-    case kAtLeastOneBrandonMichel:
-      return AtLeastOneBrandonMichelCut(univ);
 
     // If a michel's pion fails the LLR cut, remove it from the michels
     case kLLR: {
@@ -350,8 +322,6 @@ bool MinosChargeCut(const CVUniverse& univ) {
   return univ.GetDouble("MasterAnaDev_minos_trk_qp") < 0.0;
 }
 
-bool ThetamuCut(const CVUniverse& univ) { return univ.GetThetamu() < 0.3491; }
-bool DeadTimeCut(const CVUniverse& univ) { return univ.GetInt("tdead") <= 1; }
 bool WexpCut(const CVUniverse& univ, SignalDefinition signal_definition) {
   switch (signal_definition) {
     case kOnePi:
@@ -527,134 +497,6 @@ bool PmuCut(const CVUniverse& univ) {
 }
 
 //==============================================================================
-// Retired
-//==============================================================================
-// Brandon's Minos Cut
-bool BrandonMinosChargeCut(CVUniverse& univ) {
-  return univ.GetDouble("MasterAnaDev_muon_qpqpe") < 0.0;
-}
-
-// CCInclusive-style minos charge cut
-bool CCIncMinosChargeCut(CVUniverse& univ) {
-  if (univ.GetBool("MasterAnaDev_minos_used_curvature"))
-    return 1. / univ.GetDouble("MasterAnaDev_minos_trk_eqp_qp") < -5.0;
-  else if (univ.GetBool("MasterAnaDev_minos_used_range"))
-    return univ.GetBool("MasterAnaDev_minos_trk_qp") < 0.0;
-  else
-    return false;
-}
-
-bool MinosCoilCut(const CVUniverse& univ) {
-  const double MINOS_COIL_RADIUS = 210;  // mm
-  const double MAX_MINOS_RADIUS = 2500;  // mm
-  const double coilXPos = 1219.0;
-  const double coilYPos = 393.0;
-  const double minos_x =
-      univ.GetDouble("MasterAnaDev_minos_trk_end_x") + coilXPos;
-  const double minos_y =
-      univ.GetDouble("MasterAnaDev_minos_trk_end_y") + coilYPos;
-  double minosR = sqrt(pow(minos_x, 2) + pow(minos_y, 2));
-  // if (!((pow(minos_x,2) + pow(minos_y,2) )>= pow(MINOS_COIL_RADIUS, 2)) )
-  //  cout << minos_x << " " << minos_y << " " << MINOS_COIL_RADIUS << endl;
-  return (minosR > MINOS_COIL_RADIUS && minosR < MAX_MINOS_RADIUS);
-}
-
-// no isolated blobs allowed
-bool IsoBlobCut(const CVUniverse& univ) {
-  return univ.GetInt("n_iso_blob_prongs") <
-         1;  // RecoilUtils::createIsoBlobProngs
-}
-
-// IsoProngSeparation
-bool IsoProngSepCut(const CVUniverse& univ) {
-  return univ.GetLargestIsoProngSep() < 300;
-}
-
-bool AtLeastOneBrandonMichelCut(const CVUniverse& univ) {
-  // Get quality hadron candidates
-  std::vector<int> pion_candidate_indices =
-      GetQualityPionCandidateIndices(univ);
-
-  // Loop quality hadron candidates to see if they have a good brandon michel
-  for (auto pion_candidate_idx : pion_candidate_indices) {
-    int michel_views = univ.GetVecElem("MasterAnaDev_hadron_endMichel_category",
-                                       pion_candidate_idx);
-    int michel_ndigits = univ.GetVecElem(
-        "MasterAnaDev_hadron_endMichel_ndigits", pion_candidate_idx);
-    double michel_energy =
-        univ.GetVecElem("MasterAnaDev_hadron_endMichel_energy",
-                        pion_candidate_idx);  /// TODO sys universe function
-    double michel_slice_energy = univ.GetVecElem(
-        "MasterAnaDev_hadron_endMichel_slice_energy", pion_candidate_idx);
-
-    if (michel_views < 1)
-      continue;                    // no michel
-    else if (michel_views == 1) {  // 1 view
-      if (michel_energy < 55.0 && michel_ndigits < 35 &&
-          michel_slice_energy < 100.0)
-        return true;
-      else
-        continue;
-    } else if (michel_views > 1) {  // 2+3 views
-      if (michel_energy < 55.0 && michel_ndigits < 35 &&
-          michel_ndigits >= michel_views)
-        return true;
-      else
-        continue;
-    }
-  }
-  return false;
-}
-
-// ntrk are tracks not including the muon prong
-bool AtLeastOneAnchoredProngCut(const CVUniverse& univ) {
-  int ntrk = univ.GetInt("n_anchored_long_trk_prongs") +
-             univ.GetInt("n_anchored_short_trk_prongs");
-  return ntrk > 0;
-}
-
-bool ExactlyOneEndpointMichelCut(const CVUniverse& univ,
-                                 SignalDefinition signal_definition) {
-  if (signal_definition == kNPi || signal_definition == kNPiNoW) {
-    return true;
-  } else if (signal_definition == kOnePi || signal_definition == kOnePiNoW) {
-    MichelMap mm = GetQualityMichels(univ);
-    if (mm.size() == 1) {  // require only one michel
-      Michel m = (mm.begin())->second;
-      if (m.vtx == 0)
-        return false;  // so this has a no-vertex michel cut baked in
-      // pion_candidate_idx = m.vtx - 1;   // SELECT OUR PION
-      return true;
-    } else
-      return false;
-  } else {
-    std::cout << "ExactlyOneEndpointMichelcut SIGNAL DEFINITION ERROR"
-              << std::endl;
-    return false;
-  }
-}
-
-bool AtLeastOneLLRCandidateCut(const CVUniverse& univ) {
-  // Get quality hadron candidates
-  std::vector<int> pion_candidate_indices =
-      GetQualityPionCandidateIndices(univ);
-  for (auto pion_candidate_idx : pion_candidate_indices) {
-    if (LLRCut(univ, pion_candidate_idx)) return true;
-  }
-  return false;
-}
-
-bool AtLeastOneNodeCandidateCut(const CVUniverse& univ) {
-  // Get quality hadron candidates
-  std::vector<int> pion_candidate_indices =
-      GetQualityPionCandidateIndices(univ);
-  for (auto pion_candidate_idx : pion_candidate_indices) {
-    if (NodeCut(univ, pion_candidate_idx)) return true;
-  }
-  return false;
-}
-
-//==============================================================================
 // Cut Names
 //==============================================================================
 std::string GetCutName(ECuts cut) {
@@ -686,71 +528,29 @@ std::string GetCutName(ECuts cut) {
     case kMinosCharge:
       return "MINOS Charge";
 
-    case kMinosCoil:
-      return "MINOS Coil";
-
     case kMinosMuon:
       return "MINOS Muon";
-
-    case kThetaMu:
-      return "Muon Angle";
-
-    case kDeadTime:
-      return "Dead Time";
 
     case kWexp:
       return "$W_{experimental}$";
 
-    case kIsoBlobs:
-      return "$<$1 Isolated Blobs";
-
     case kIsoProngs:
       return "$<$2 Isolated Prongs";
-
-    case kIsoProngSep:
-      return "Iso Prong Sep $<$ 300";
-
-    case kNProngs:
-      return "Max 2 Hadr Prongs";
 
     case kNPionCandidates:
       return "$\\pi$ candidate";
 
-    case kPionCandidateQuality:
-      return "Quality $\\pi$ candidate";
-
     case kAtLeastOneMichel:
       return "$>$= 1 Michel";
 
-    case kAtLeastOneBrandonMichel:
-      return "$>$= 1 Brandon Michel";
-
-    case kAtLeastOneAnchoredProng:
-      return "$>$= 1 Anchored Prong";
-
     case kAtLeastOnePionCandidateTrack:
       return "$>$= 1 Hadron Track";
-
-    case kAtLeastOneLLRCandidate:
-      return "$>$= 1 Track Pass LLR Cut";
-
-    case kAtLeastOneNodeCandidate:
-      return "$>$= 1 Track Pass Node Cut";
-
-    case kExactlyOneEndpointMichel:
-      return "== 1 Michel";
 
     case kNode:
       return "Node";
 
     case kPionMult:
       return "Pion Multiplicity";
-
-    case kOldMichel:
-      return "Old Michel Cut";
-
-    case kdEdx:
-      return "dEdx PID";
 
     case kLLR:
       return "LLR PID";
